requirements.txt
gradio
langchain
langchain-google-genai
chromadb
python-dotenv
langchain-chroma